Week 2 HW
How difficult did you find this (out of 10)? 5;
Was there anything that you struggled with? I need to practice more on the flexbox and how to differ its properties;
Is there anything that you'd like some further information on? Maybe more examples with comparisons of the most used properties in flexbox;
Roughly how long did it take? Probably a cumulative of 5 to 6 hours on the homeworks