Final Project Phase 1
How difficult did you find this (out of 10)? 7;\nWas there anything that you struggled with? Everything was clear. The struggle was with structuring the idea and implementing the first phase plus flexbox and Grid;\nIs there anything that you'd like some further information on? An extra session clarifying the implementations of flexbox in grid in CSS. It's still not clear enough; Roughly how long did it take? Cumulative of 4-5 hours